.activities-organizer/requirements.txt
Jinja2>=3.1.2
Pillow>=10.1.0
html2image>=2.0.4.3